backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
redis==5.2.1
anthropic==0.40.0
twilio==9.4.0
resend==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
apscheduler==3.10.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
python-dotenv==1.0.1
